% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allR4WSItasktest}
\alias{allR4WSItasktest}
\title{Example R4WSI Tasks with Regular and Gold-Standard Tasks}
\format{
A data frame of 20 rows and 6 columns.
\describe{
\item{\code{topic}}{Index of topics}
\item{\code{id}}{Index of topics}
\item{\code{doc}}{Example documents associated with each topic}
\item{\code{opt1}}{Words set option 1}
\item{\code{opt2}}{Words set option 2}
\item{\code{opt3}}{Words set option 3}
\item{\code{optcrt}}{Words set option 4, also the correct choice}
}
}
\usage{
data(allR4WSItasktest)
}
\description{
Data frame of 20 example R4WSI0 Tasks, with 5 of them being gold-standard and 15 of them not.
}
\keyword{datasets}
