% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAgree.R
\name{checkAgree}
\alias{checkAgree}
\title{Check Agreement Rate between Identical Trails}
\usage{
checkAgree(results1, results2, key, type = NULL)
}
\arguments{
\item{results1}{first batch of results; outputs from getResults()}

\item{results2}{first batch of results; outputs from getResults()}

\item{key}{the local task record; outputs from recordTasks()}

\item{type}{Task structures to be specified. Must be one of "WI" (word intrusion),
"T8WSI" (top 8 word set intrusion), "R4WSI" (random 4 word set intrusion),
"LI" (Label Intrusion), and "OL" (Optimal Label)}
}
\description{
Check Agreement Rate between Identical Trails
}
\details{
Evaluate workers' performance by agreement rate between identical trails
(Notice that this means the two input, results1 and results2, must be identical.);
Return 1) the exact agreement rate when both workers agree on the exact same choice, and
2) the binary agreement rate when both workers get the task either right or wrong simultaneously
}
