% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combMass.R
\name{combMass}
\alias{combMass}
\title{Combine the mass of words with the same root}
\usage{
combMass(mod = NULL, vocab = NULL, beta = NULL)
}
\arguments{
\item{mod}{Fitted structural topic models.}

\item{vocab}{A character vector specifying the words in the corpus. Usually, it
can be found in topic model output.}

\item{beta}{A matrix of word probabilities for each topic. Each row represents a
topic and each column represents a word. Note this should not be in the logged form.}
}
\description{
Combine the mass of words with the same root
}
\details{
Use as a preparing step for validating unstemmed topic models.
}
