% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalResults.R
\name{evalResults}
\alias{evalResults}
\title{Evaluate results}
\usage{
evalResults(results, key, type = NULL)
}
\arguments{
\item{results}{results of human choice; outputs from getResults()}

\item{key}{the local task record; outputs form recordTasks()}

\item{type}{Task structures to be specified. Must be one of "WI" (word intrusion),
"T8WSI" (top 8 word set intrusion), "R4WSI" (random 4 word set intrusion),
"LI" (Label Intrusion), and "OL" (Optimal Label)}
}
\description{
Evaluate results
}
\details{
Evaludate worker performance by gold-standard HITs;
Return the accuracy rate (proportion correct) for a specified batch
}
