% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goldR4WSItest}
\alias{goldR4WSItest}
\title{Example Gold-Standard R4WSI0 Tasks}
\format{
A data frame of 5 rows and 6 columns.
\describe{
\item{\code{topic}}{Index of topics}
\item{\code{doc}}{Example documents associated with each topic}
\item{\code{opt1}}{Words set option 1}
\item{\code{opt2}}{Words set option 2}
\item{\code{opt3}}{Words set option 3}
\item{\code{optcrt}}{Words set option 4, also the correct choice}
}
}
\usage{
data(goldR4WSItest)
}
\description{
Data frame of 5 example gold-standard R4WSI0 Tasks.
}
\keyword{datasets}
