% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heldouttest}
\alias{heldouttest}
\title{An Example Heldout Test Set}
\format{
A list of the heldout documents, vocab, and missing.
}
\source{
See \url{https://CRAN.R-project.org/package=stm} for more details.
}
\usage{
data(heldouttest)
}
\description{
An output from the \code{make.heldout} function of the \code{stm} package.
}
\references{
Roberts, Margaret E., Brandon M. Stewart, and Dustin Tingley. "Stm: An R package for structural
 topic models." Journal of Statistical Software 91 (2019): 1-40.
}
