% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keypostedtest}
\alias{keypostedtest}
\title{Example Answer Keys}
\format{
A list of two data frames. Similar to \code{recordtest}.

\describe{
\item{\code{data.frame1}}{A data frame of tasks with the \code{optcrt} indicating the 
machine predicted choice.}
\item{\code{data.frame2}}{A data frame of tasks with randomized choices. 
Exactly the same with what would be sent online.}
}
}
\usage{
data(keypostedtest)
}
\description{
Example Answer Keys
}
