% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masstest}
\alias{masstest}
\title{An Example of the Combined Mass for Words with the Same Roots}
\format{
A list of two.
}
\usage{
data(masstest)
}
\description{
A list of two with the words (the most frequent form in each topic) and the corresponding word 
probabilities.
}
\details{
\describe{
\item{\code{vocab}}{A matrix of words for each topic. Each row represents a
topic and each column represents the words. Words with the same roots are
only represented by the most common form in that topic.}
\item{\code{beta}}{A matrix of combined word probabilities for each topic. 
Each row represents a topic and each column represents a combined word.}
}
}
