% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixGold.R
\name{mixGold}
\alias{mixGold}
\title{Mix the gold-standard tasks with the tasks need to be validated}
\usage{
mixGold(tasks, golds)
}
\arguments{
\item{tasks}{All tasks need to be validated}

\item{golds}{Gold standard tasks with the same structure}
}
\value{
A data frame with the same structure as the input, where gold-standard tasks are randomly inserted
}
\description{
Mix the gold-standard tasks with the tasks need to be validated
}
