% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modtest}
\alias{modtest}
\title{An Example Topic Model}
\format{
A STM object.
}
\source{
See \url{https://CRAN.R-project.org/package=stm} for more details.
}
\usage{
data(modtest)
}
\description{
A structural topic model (STM) object generated from the \code{stm} package using a random 
sample of US senators' Facebook posts.
}
\references{
Roberts, Margaret E., Brandon M. Stewart, and Dustin Tingley. "Stm: An R package for structural
 topic models." Journal of Statistical Software 91 (2019): 1-40.
}
