% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R4WSItasktest}
\alias{R4WSItasktest}
\title{Example R4WSI0 Tasks}
\format{
A matrix with 15 rows and 6 columns.
\describe{
\item{\code{topic}}{Index of topics}
\item{\code{doc}}{Example documents associated with each topic}
\item{\code{opt1}}{Words set option 1}
\item{\code{opt2}}{Words set option 2}
\item{\code{opt3}}{Words set option 3}
\item{\code{optcrt}}{Words set option 4, also the correct choice}
}
}
\usage{
data(R4WSItasktest)
}
\description{
Data of 15 example R4WSI0 Tasks structured as a matrix.
}
\details{
Please note that the difference between the R4WSI0 examples used here and the R4WSI tasks
is that the R4WSI tasks do not present any documents.
}
\keyword{datasets}
