% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordTasks.R
\name{recordTasks}
\alias{recordTasks}
\title{Reform tasks to facilitate sending to Mturk}
\usage{
recordTasks(type, tasks, path)
}
\arguments{
\item{type}{one of WI, T8WSI, R4WSI}

\item{tasks}{outputs from validateTopic(), validateLabel(), or mixGold() if users mix in gold-standard HITs}

\item{path}{path to record the tasks (with meta-information)}
}
\description{
Reform tasks to facilitate sending to Mturk
}
\details{
Randomize the order of options and record the tasks in a specified local directory
}
