% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recordtest}
\alias{recordtest}
\title{Example Local Record of the R4WSI Tasks}
\format{
A list of two data frames.
\describe{
\item{\code{data.frame1}}{A data frame of tasks with the \code{optcrt} indicating the 
machine preficted choice.}
\item{\code{data.frame2}}{A data frame of tasks with randomized choices. 
Exactly the same with what would be sent online.}
}
}
\usage{
data(recordtest)
}
\description{
Local record generated by the \code{recordTasks} function.
}
\details{
To be compared with the answers from the online workers to evaluate the topic model performance.
}
