% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultstest}
\alias{resultstest}
\title{Example Results Retrieved from Mturk}
\format{
A data frame of ten example tasks retrieved from the Mturk with or without 
online workers' answers.

\describe{
\item{\code{assignment_id}}{Assignment id. Mturk assigned. If 0, then the task hasn't been completed.}
\item{\code{batch_id}}{User specified batch id.}
\item{\code{completed_at}}{Timestamp when the task was completed. If 0, then the task hasn't been completed.}
\item{\code{local_task_id}}{Local task id.}
\item{\code{mturk_hit_id}}{Mturk HIT id. Mturk assigned.}
\item{\code{result}}{Choice made by the worker. 1-4. If 0, then the task hasn't been completed.}
\item{\code{worker_id}}{Mturk worker id. If 0, then the task hasn't been completed.}
}
}
\usage{
data(resultstest)
}
\description{
Example Results Retrieved from Mturk
}
\keyword{datasets}
