% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stmPreptest}
\alias{stmPreptest}
\title{An Example Object of Prepared Documents}
\format{
A list containing a documents and vocab object.
}
\source{
See \url{https://CRAN.R-project.org/package=stm} for more details.
}
\usage{
data(stmPreptest)
}
\description{
An output from the \code{prepDocuments} function of the \code{stm} package.
}
\references{
Roberts, Margaret E., Brandon M. Stewart, and Dustin Tingley. "Stm: An R package for structural
 topic models." Journal of Statistical Software 91 (2019): 1-40.
}
